Crate axmem

Address space of user mode management of microkernel version of ArceOS.

Structs

AddrSpace

A user address space

AddrSpaceInner

mutable inner objects of address space

Constants

MMAP_AREA_END

end of default mmap areas

MMAP_AREA_START

start of default mmap areas

TRAMPOLINE_START

address of trampoline page

USER_START

Start of binary

USTACK_SIZE

size of user stack

USTACK_START

top of user stack

Traits

CurrentAddrSpace

interface to get current process’s address space

Functions

alloc_user_page

alloc a user space

copy_byte_buffer_to_user

Copy a u8 array data' from current memory space into position ptr’ of the userspace `token’(not used)

copy_data_to_user

Copy a data' with type T’ from current memory space into position ptr' of the userspace token’ (not used)

copy_slice_from_user

copy a slice of current process into kernel address space

copy_str_from_user

copy a str in current process into kernel address space

get_satp

get root of page table of current process

global_sbrk

manage heap space of current process

mmap_page

mmap a page of current process

munmap_page

unmap a page of current process

translate_addr

translate a virtual address to phyical address of current process

translate_buffer

translate a slice from current process