Struct axmem::MapArea

pub struct MapArea {
    pub pages: Vec<Option<PhysPage>>,
    pub vaddr: VirtAddr,
    pub flags: MappingFlags,
    pub backend: Option<MemBackend>,
}

A continuous virtual area in user memory.

NOTE: Cloning a MapArea needs allocating new phys pages and modifying a page table. So Clone trait won’t implemented.

Fields
pages: Vec<Option<PhysPage>>

phys pages of this area

vaddr: VirtAddr

start virtual address

flags: MappingFlags

mapping flags of this area

backend: Option<MemBackend>

whether the area is backed by a file

Implementations

impl MapArea

pub fn new_lazy( start: VirtAddr, num_pages: usize, flags: MappingFlags, backend: Option<MemBackend>, page_table: &mut PageTable ) -> Self

Create a lazy-load area and map it in page table (page fault PTE).

pub fn new_alloc( start: VirtAddr, num_pages: usize, flags: MappingFlags, data: Option<&[u8]>, backend: Option<MemBackend>, page_table: &mut PageTable ) -> AxResult<Self>

Allocated an area and map it in page table.

pub fn dealloc(&mut self, page_table: &mut PageTable)

Deallocate all phys pages and unmap the area in page table.

pub fn handle_page_fault( &mut self, addr: VirtAddr, flags: MappingFlags, page_table: &mut PageTable ) -> bool

如果处理失败，返回false，此时直接退出当前程序

pub fn sync_page_with_backend(&mut self, page_index: usize)

Sync pages in index back to self.backend (if there is one).

Panics

Panics if index is out of bounds.

pub fn shrink_left(&mut self, new_start: VirtAddr, page_table: &mut PageTable)

Deallocate some pages from the start of the area. This function will unmap them in a page table. You need to flush TLB after this function.

pub fn shrink_right(&mut self, new_end: VirtAddr, page_table: &mut PageTable)

Deallocate some pages from the end of the area. This function will unmap them in a page table. You need to flush TLB after this function.

pub fn split(&mut self, addr: VirtAddr) -> Self

Split this area into 2.

pub fn split3(&mut self, start: VirtAddr, end: VirtAddr) -> (Self, Self)

Split this area into 3.

pub fn remove_mid( &mut self, left_end: VirtAddr, right_start: VirtAddr, page_table: &mut PageTable ) -> Self

Create a second area in the right part of the area, [self.vaddr, left_end) and [right_start, self.end_va()). This function will unmap deleted pages in a page table. You need to flush TLB after calling this.

impl MapArea

pub fn size(&self) -> usize

return the size of the area, which thinks the page size is default 4K.

pub fn end_va(&self) -> VirtAddr

return the end virtual address of the area.

pub fn allocated(&self) -> bool

return whether all the pages have been allocated.

pub unsafe fn as_slice(&self) -> &[u8]

Safety

This function is unsafe because it dereferences a raw pointer. It will return a slice of the area’s memory, whose len is the same as the area’s size.

pub fn fill(&mut self, byte: u8)

Fill self with byte.

pub fn overlap_with(&self, start: VirtAddr, end: VirtAddr) -> bool

If [start, end) overlaps with self.

pub fn contained_in(&self, start: VirtAddr, end: VirtAddr) -> bool

If [start, end] contains self.

pub fn contains(&self, start: VirtAddr, end: VirtAddr) -> bool

If self contains [start, end].

pub fn strict_contain(&self, start: VirtAddr, end: VirtAddr) -> bool

If self strictly contains [start, end], which stands for the start and end are not equal to self’s.

pub fn update_flags(&mut self, flags: MappingFlags, page_table: &mut PageTable)

Update area’s mapping flags and write it to page table. You need to flush TLB after calling this function.

pub fn clone_alloc(&self, page_table: &mut PageTable) -> AxResult<Self>

Allocating new phys pages and clone it self. This function will modify the page table as well.