Struct axmem::MemorySet

pub struct MemorySet { /* private fields */ }

PageTable + MemoryArea for a process (task)

Implementations

impl MemorySet

pub fn page_table_token(&self) -> usize

Get the root page table token.

pub fn new_empty() -> Self

Create a new empty MemorySet.

pub fn new_memory_set() -> Self

Create a new MemorySet

pub fn page_table_root_ppn(&self) -> PhysAddr

The root page table physical address.

pub fn max_va(&self) -> VirtAddr

The max virtual address of the areas in this memory set.

pub fn new_region( &mut self, vaddr: VirtAddr, size: usize, shared: bool, flags: MappingFlags, data: Option<&[u8]>, backend: Option<MemBackend> )

Allocate contiguous region. If no data, it will create a lazy load region.

pub fn split_for_area(&mut self, start: VirtAddr, size: usize)

Make [start, end) unmapped and dealloced. You need to flush TLB after this.

NOTE: modified map area will have the same PhysAddr.

pub fn find_free_area(&self, hint: VirtAddr, size: usize) -> Option<VirtAddr>

Find a free area with given start virtual address and size. Return the start address of the area.

pub fn mmap( &mut self, start: VirtAddr, size: usize, flags: MappingFlags, shared: bool, fixed: bool, backend: Option<MemBackend> ) -> AxResult<usize>

mmap. You need to flush tlb after this.

pub fn munmap(&mut self, start: VirtAddr, size: usize)

munmap. You need to flush TLB after this.

pub fn msync(&mut self, start: VirtAddr, size: usize)

msync

pub fn mprotect(&mut self, start: VirtAddr, size: usize, flags: MappingFlags)

Edit the page table to update flags in given virt address segment. You need to flush TLB after calling this function.

NOTE: It’s possible that this function will break map areas into two for different mapping flag settings.

pub fn handle_page_fault( &mut self, addr: VirtAddr, flags: MappingFlags ) -> AxResult<()>

It will map newly allocated page in the page table. You need to flush TLB after this.

pub fn unmap_user_areas(&mut self)

将用户分配的页面从页表中直接解映射，内核分配的页面依然保留

pub fn query( &self, vaddr: VirtAddr ) -> AxResult<(PhysAddr, MappingFlags, PageSize)>

Query the page table to get the physical address, flags and page size of the given virtual

pub fn map_page_without_alloc( &mut self, vaddr: VirtAddr, paddr: PhysAddr, flags: MappingFlags ) -> AxResult<()>

Map a 4K region without allocating physical memory.

pub fn create_shared_mem( key: i32, size: usize, pid: u64, uid: u32, gid: u32, mode: u16 ) -> AxResult<(i32, SharedMem)>

Create a new SharedMem with given key. You need to add the returned SharedMem to global SHARED_MEMS or process’s private_mem.

Panics: SharedMem with the key already exist.

pub fn add_shared_mem(shmid: i32, mem: SharedMem)

Panics: shmid is already taken.

pub fn add_private_shared_mem(&mut self, shmid: i32, mem: SharedMem)

Panics: shmid is already taken in the process.

pub fn get_shared_mem(shmid: i32) -> Option<Arc<SharedMem>>

Get a SharedMem by shmid.

pub fn get_private_shared_mem(&self, shmid: i32) -> Option<Arc<SharedMem>>

Get a private SharedMem by shmid.

pub fn attach_shared_mem( &mut self, mem: Arc<SharedMem>, addr: VirtAddr, flags: MappingFlags )

Attach a SharedMem to the memory set.

pub fn detach_shared_mem(&mut self, _shmid: i32)

Detach a SharedMem from the memory set.

TODO: implement this

pub fn mremap( &mut self, old_start: VirtAddr, old_size: usize, new_size: usize ) -> isize

mremap: change the size of a mapping, potentially moving it at the same time.

impl MemorySet

pub fn manual_alloc_for_lazy(&mut self, addr: VirtAddr) -> AxResult<()>

判断某一个虚拟地址是否在内存集中。 若当前虚拟地址在内存集中，且对应的是lazy分配，暂未分配物理页的情况下， 则为其分配物理页面。

若不在内存集中，则返回None。

若在内存集中，且已经分配了物理页面，则不做处理。

pub fn manual_alloc_range_for_lazy( &mut self, start: VirtAddr, end: VirtAddr ) -> AxResult<()>

暴力实现区间强制分配 传入区间左闭右闭

pub fn manual_alloc_type_for_lazy<T: Sized>( &mut self, obj: *const T ) -> AxResult<()>

判断某一个类型的某一个对象是否被分配

impl MemorySet

pub fn clone_or_err(&mut self) -> AxResult<Self>

Clone the MemorySet. This will create a new page table and map all the regions in the old page table to the new one.

If it occurs error, the new MemorySet will be dropped and return the error.