Crate axmem

The memory management module, which implements the memory space management of the process.

Structs

MapArea

A continuous virtual area in user memory.

MemBackend

File backend for Lazy load MapArea. file should be a file holding a offset value. Normally, MemBackend won’t share a file with other things, so we use a Box here.

MemorySet

PageTable + MemoryArea for a process (task)

Statics

KEY_TO_SHMID

The map from key to shmid. It’s used to query shmid from key.

SHARED_MEMS

This struct only hold SharedMem that are not IPC_PRIVATE. IPC_PRIVATE SharedMem will be stored in MemorySet::detached_mem.

Functions

check_page_table_entry_validity

验证地址是否已分配页面